대규모 namespace 이사. Behavior의 개념 변경. 이제 Behavior는 직접 무언가를 실행한다기보다는 진짜 실행을 위한 Command를 생성하는 역할을 함.

+++ GfEngine/Models/Items/Weapon.cs
@@ -5,7 +5,7 @@ namespace GfEngine.Models.Items
 {
 	public class Weapon : Item
 	{
-		public int Power { get; set; }
+		public float Power { get; set; } // 공격시 공격력에 곱해질 계수
 		public WeaponType Type { get; set; }
 		public BuffSet EquipBuff { get; set; }
 		public HashSet<WeaponTag> WTags { get; set; }
